# cedar-terminology-server-application/src/main/java/org/metadatacenter/cedar/terminology/resources/bioportal/OntologyResource.java
package org.metadatacenter.cedar.terminology.resources.bioportal;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.metadatacenter.util.http.CedarError;
import org.metadatacenter.cedar.cache.Cache;
import org.metadatacenter.cedar.terminology.resources.AbstractTerminologyServerResource;
import org.metadatacenter.config.CedarConfig;
import org.metadatacenter.exception.CedarException;
import org.metadatacenter.rest.context.CedarRequestContext;
import org.metadatacenter.rest.exception.CedarAssertionException;
import org.metadatacenter.terms.domainObjects.Ontology;
import org.metadatacenter.terms.domainObjects.OntologyClass;
import org.metadatacenter.terms.domainObjects.OntologyVersion;
import org.metadatacenter.terms.domainObjects.VersionDiff;
import org.metadatacenter.terms.domainObjects.VersionTriple;
import org.metadatacenter.terms.domainObjects.OntologyProperty;
import org.metadatacenter.util.http.CedarResponse;
import org.metadatacenter.util.json.JsonMapper;

import jakarta.ws.rs.GET;
import jakarta.ws.rs.Path;
import jakarta.ws.rs.PathParam;
import jakarta.ws.rs.Produces;
import jakarta.ws.rs.core.MediaType;
import jakarta.ws.rs.core.Response;
import javax.xml.ws.http.HTTPException;
import java.io.IOException;
import java.util.ArrayList;
import java.util.List;
import java.util.concurrent.ExecutionException;

import static org.metadatacenter.rest.assertion.GenericAssertions.LoggedIn;

@Path("/bioportal")
@Produces(MediaType.APPLICATION_JSON)
@Tag(name = "Ontologies")
@SecurityRequirement(name = "api_key")
public class OntologyResource extends AbstractTerminologyServerResource {

  public OntologyResource(CedarConfig cedarConfig) {
    super(cedarConfig);
  }

  @GET
  @Path("ontologies")
  @Operation(summary = "Find all ontologies", description = "Find all ontologies.")
  @ApiResponses({
      @ApiResponse(responseCode = "200", description = "Every ontology the server serves", content = @Content(array = @ArraySchema(schema = @Schema(implementation = Ontology.class)))),
      @ApiResponse(responseCode = "400", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Bad request"),
      @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Unauthorized"),
      @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Forbidden"),
      @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Not found"),
      @ApiResponse(responseCode = "500", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Internal server error")
  })
  public Response findAllOntologies() throws CedarException {
    CedarRequestContext ctx = buildRequestContext();
    ctx.must(ctx.user()).be(LoggedIn);
    try {
      List<Ontology> ontologies = new ArrayList<>(Cache.getOntologies().values());
      return Response.ok().entity(JsonMapper.MAPPER.valueToTree(ontologies)).build();
    } catch (HTTPException e) {
      return relayedBioPortalFailure(e);
    } catch (ExecutionException e) {
      throw new CedarAssertionException(e);
    }
  }

  @GET
  @Path("ontologies/{id}")
  @Operation(summary = "Find ontology by id", description = "Find ontology by id.")
  @ApiResponses({
      @ApiResponse(responseCode = "200", description = "The ontology", content = @Content(schema = @Schema(implementation = Ontology.class))),
      @ApiResponse(responseCode = "400", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Bad request"),
      @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Unauthorized"),
      @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Forbidden"),
      @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Not found"),
      @ApiResponse(responseCode = "500", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Internal server error")
  })
  public Response findOntology(
      @Parameter(description = "BioPortal ontology identifier. Examples: NCIT, FMA, OBI.", required = true)
      @PathParam("id") String id) throws CedarException {
    CedarRequestContext ctx = buildRequestContext();
    ctx.must(ctx.user()).be(LoggedIn);
    try {
      Ontology ontologies = Cache.getOntology(id);
      if (ontologies == null) {
        return CedarResponse.notFound().build();
      }
      return Response.ok().entity(JsonMapper.MAPPER.valueToTree(ontologies)).build();
    } catch (HTTPException e) {
      return relayedBioPortalFailure(e);
    } catch (ExecutionException e) {
      throw new CedarAssertionException(e);
    }
  }

  @GET
  @Path("ontologies/{id}/versions")
  @Operation(summary = "List local versions of an ontology",
      description = "Versions of an ontology available in the local, version-pinned store, each with "
          + "its content-hash id, self-declared version, release timestamp, effectiveDate (the "
          + "release day, or ingest day when the source records no release), and whether it is the "
          + "current one. Empty when the ontology is served from BioPortal (which has no equivalent).",
      tags = {"Ontologies"})
  @ApiResponses({
      @ApiResponse(responseCode = "200", description = "The versions held locally, newest state marked latest", content = @Content(array = @ArraySchema(schema = @Schema(implementation = OntologyVersion.class)))),
      @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Unauthorized"),
      @ApiResponse(responseCode = "500", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Internal server error")
  })
  public Response getVersions(
      @Parameter(description = "Ontology acronym. Examples: DOID, INCENTIVE.", required = true)
      @PathParam("id") String id) throws CedarException {
    CedarRequestContext ctx = buildRequestContext();
    ctx.must(ctx.user()).be(LoggedIn);
    try {
      List<OntologyVersion> versions = terminologyService.getVersions(id);
      return Response.ok().entity(JsonMapper.MAPPER.valueToTree(versions)).build();
    } catch (IOException e) {
      throw new CedarAssertionException(e);
    }
  }

  @GET
  @Path("ontologies/{id}/versions/current")
  @Operation(summary = "Resolve the current version triple of an ontology",
      description = "The version triple {id (content hash), effectiveDate, declaredVersion} of the "
          + "ontology's current (latest) local snapshot. This is the publish-time 'resolve current' "
          + "capability: a publish pipeline calls it per value-constraint entry to freeze the entry "
          + "against its ontology's current state. 404 when the ontology is not served locally "
          + "(BioPortal has no content-hash triple to freeze).",
      tags = {"Ontologies"})
  @ApiResponses({
      @ApiResponse(responseCode = "200", description = "The current version triple", content = @Content(schema = @Schema(implementation = VersionTriple.class))),
      @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Unauthorized"),
      @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Ontology not served locally"),
      @ApiResponse(responseCode = "500", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Internal server error")
  })
  public Response resolveCurrentVersion(
      @Parameter(description = "Ontology acronym. Examples: DOID, INCENTIVE.", required = true)
      @PathParam("id") String id) throws CedarException {
    CedarRequestContext ctx = buildRequestContext();
    ctx.must(ctx.user()).be(LoggedIn);
    try {
      VersionTriple triple = terminologyService.resolveCurrentVersion(id);
      if (triple == null) {
        return CedarResponse.notFound().build();
      }
      return Response.ok().entity(JsonMapper.MAPPER.valueToTree(triple)).build();
    } catch (IOException e) {
      throw new CedarAssertionException(e);
    }
  }

  @GET
  @Path("classes/version-current")
  @Operation(summary = "Resolve the current version triple for a class IRI",
      description = "The version triple of the ontology that owns the given class/term IRI — the "
          + "freeze-on-publish capability for a class-valued constraint, which names a term but not "
          + "its ontology. The IRI's namespace is mapped to its ontology, then that ontology's current "
          + "triple is returned. 404 when the ontology cannot be determined unambiguously or is not "
          + "served locally.",
      tags = {"Classes"})
  @ApiResponses({
      @ApiResponse(responseCode = "200", description = "The version triple of the ontology that owns the class", content = @Content(schema = @Schema(implementation = VersionTriple.class))),
      @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Unauthorized"),
      @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Ontology for the class not resolvable locally"),
      @ApiResponse(responseCode = "500", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Internal server error")
  })
  public Response resolveCurrentVersionForClass(
      @Parameter(description = "Class/term IRI. Example: http://purl.obolibrary.org/obo/DOID_9351.",
          required = true)
      @jakarta.ws.rs.QueryParam("uri") String uri) throws CedarException {
    CedarRequestContext ctx = buildRequestContext();
    ctx.must(ctx.user()).be(LoggedIn);
    try {
      VersionTriple triple = terminologyService.resolveCurrentVersionForClass(uri);
      if (triple == null) {
        return CedarResponse.notFound().build();
      }
      return Response.ok().entity(JsonMapper.MAPPER.valueToTree(triple)).build();
    } catch (IOException e) {
      throw new CedarAssertionException(e);
    }
  }

  @GET
  @Path("ontologies/{id}/versions/diff")
  @Operation(summary = "Diff two local versions of an ontology",
      description = "The vocabulary diff (concept and subsumption-edge additions/removals, newly "
          + "obsoleted, capped concept samples) between two versions in the local store, given by "
          + "version_id or tag. 404 when the ontology or a version is not served locally.",
      tags = {"Ontologies"})
  @ApiResponses({
      @ApiResponse(responseCode = "200", description = "The vocabulary diff between the two versions", content = @Content(schema = @Schema(implementation = VersionDiff.class))),
      @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Unauthorized"),
      @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Ontology or version not found locally"),
      @ApiResponse(responseCode = "500", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Internal server error")
  })
  public Response diffVersions(
      @Parameter(description = "Ontology acronym.", required = true) @PathParam("id") String id,
      @Parameter(description = "Base version (version_id or tag).", required = true)
      @jakarta.ws.rs.QueryParam("from") String from,
      @Parameter(description = "Target version (version_id or tag, e.g. latest).", required = true)
      @jakarta.ws.rs.QueryParam("to") String to) throws CedarException {
    CedarRequestContext ctx = buildRequestContext();
    ctx.must(ctx.user()).be(LoggedIn);
    try {
      var diff = terminologyService.diffVersions(id, from, to);
      if (diff == null) {
        return CedarResponse.notFound().build();
      }
      return Response.ok().entity(JsonMapper.MAPPER.valueToTree(diff)).build();
    } catch (IOException e) {
      throw new CedarAssertionException(e);
    }
  }

  @GET
  @Path("ontologies/{ontology}/classes/roots")
  @Operation(summary = "Get root classes", description = "Get root classes in a particular ontology. For the CEDARPC ontology, all provisional classes in it " +
          "will be returned.", tags = {"Classes", "Ontologies"})
  @ApiResponses({
      @ApiResponse(responseCode = "200", description = "The ontology's root classes", content = @Content(array = @ArraySchema(schema = @Schema(implementation = OntologyClass.class)))),
      @ApiResponse(responseCode = "400", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Bad request"),
      @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Unauthorized"),
      @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Forbidden"),
      @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Not found"),
      @ApiResponse(responseCode = "500", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Internal server error")
  })
  public Response findRootClasses(
      @Parameter(description = "BioPortal ontology identifier. Examples: NCIT, FMA, OBI.", required = true)
      @PathParam("ontology") String ontology) throws CedarException {
    CedarRequestContext ctx = buildRequestContext();
    ctx.must(ctx.user()).be(LoggedIn);
    try {
      boolean isFlat = Cache.isFlat(ontology);
      List<OntologyClass> roots = terminologyService.getRootClasses(ontology, isFlat, apiKey);
      return Response.ok().entity(JsonMapper.MAPPER.valueToTree(roots)).build();
    } catch (HTTPException e) {
      return relayedBioPortalFailure(e);
    } catch (IOException | ExecutionException e) {
      throw new CedarAssertionException(e);
    }
  }

  @GET
  @Path("ontologies/{ontology}/properties/roots")
  @Operation(summary = "Get root properties", description = "Get root properties in a particular ontology.", tags = {"Properties", "Ontologies"})
  @ApiResponses({
      @ApiResponse(responseCode = "200", description = "The ontology's root properties", content = @Content(array = @ArraySchema(schema = @Schema(implementation = OntologyProperty.class)))),
      @ApiResponse(responseCode = "400", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Bad request"),
      @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Unauthorized"),
      @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Forbidden"),
      @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Not found"),
      @ApiResponse(responseCode = "500", content = @Content(schema = @Schema(implementation = CedarError.class)), description = "Internal server error")
  })
  public Response findRootProperties(
      @Parameter(description = "BioPortal ontology identifier. Examples: NCIT, FMA, OBI.", required = true)
      @PathParam("ontology") String ontology) throws CedarException {
    CedarRequestContext ctx = buildRequestContext();
    ctx.must(ctx.user()).be(LoggedIn);
    try {
      List<OntologyProperty> roots = terminologyService.getRootProperties(ontology, apiKey);
      return Response.ok().entity(JsonMapper.MAPPER.valueToTree(roots)).build();
    } catch (HTTPException e) {
      return relayedBioPortalFailure(e);
    } catch (IOException e) {
      throw new CedarAssertionException(e);
    }
  }

}