# cedar-terminology-server-application/src/main/resources/assets/swagger-api/swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "description" : "CEDAR Terminology Server API",
    "title" : "CEDAR Terminology Server API",
    "version" : "2.9.10-SNAPSHOT"
  },
  "paths" : {
...
    "/bioportal/classes/version-current" : {
      "get" : {
        "description" : "The version triple of the ontology that owns the given class/term IRI — the freeze-on-publish capability for a class-valued constraint, which names a term but not its ontology. The IRI's namespace is mapped to its ontology, then that ontology's current triple is returned. 404 when the ontology cannot be determined unambiguously or is not served locally.",
        "operationId" : "resolveCurrentVersionForClass",
        "parameters" : [ {
          "description" : "Class/term IRI. Example: http://purl.obolibrary.org/obo/DOID_9351.",
          "in" : "query",
          "name" : "uri",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VersionTriple"
                }
              }
            },
            "description" : "The version triple of the ontology that owns the class"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Ontology for the class not resolvable locally"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Resolve the current version triple for a class IRI",
        "tags" : [ "Classes", "Ontologies" ]
      }
    },
...
    "/bioportal/ontologies" : {
      "get" : {
        "description" : "Find all ontologies.",
        "operationId" : "findAllOntologies",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Ontology"
                  }
                }
              }
            },
            "description" : "Every ontology the server serves"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Bad request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Not found"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Find all ontologies",
        "tags" : [ "Ontologies" ]
      }
    },
    "/bioportal/ontologies/{id}" : {
      "get" : {
        "description" : "Find ontology by id.",
        "operationId" : "findOntology",
        "parameters" : [ {
          "description" : "BioPortal ontology identifier. Examples: NCIT, FMA, OBI.",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Ontology"
                }
              }
            },
            "description" : "The ontology"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Bad request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Not found"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Find ontology by id",
        "tags" : [ "Ontologies" ]
      }
    },
    "/bioportal/ontologies/{id}/versions" : {
      "get" : {
        "description" : "Versions of an ontology available in the local, version-pinned store, each with its content-hash id, self-declared version, release timestamp, effectiveDate (the release day, or ingest day when the source records no release), and whether it is the current one. Empty when the ontology is served from BioPortal (which has no equivalent).",
        "operationId" : "getVersions",
        "parameters" : [ {
          "description" : "Ontology acronym. Examples: DOID, INCENTIVE.",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/OntologyVersion"
                  }
                }
              }
            },
            "description" : "The versions held locally, newest state marked latest"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "List local versions of an ontology",
        "tags" : [ "Ontologies" ]
      }
    },
    "/bioportal/ontologies/{id}/versions/current" : {
      "get" : {
        "description" : "The version triple {id (content hash), effectiveDate, declaredVersion} of the ontology's current (latest) local snapshot. This is the publish-time 'resolve current' capability: a publish pipeline calls it per value-constraint entry to freeze the entry against its ontology's current state. 404 when the ontology is not served locally (BioPortal has no content-hash triple to freeze).",
        "operationId" : "resolveCurrentVersion",
        "parameters" : [ {
          "description" : "Ontology acronym. Examples: DOID, INCENTIVE.",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VersionTriple"
                }
              }
            },
            "description" : "The current version triple"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Ontology not served locally"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Resolve the current version triple of an ontology",
        "tags" : [ "Ontologies" ]
      }
    },
    "/bioportal/ontologies/{id}/versions/diff" : {
      "get" : {
        "description" : "The vocabulary diff (concept and subsumption-edge additions/removals, newly obsoleted, capped concept samples) between two versions in the local store, given by version_id or tag. 404 when the ontology or a version is not served locally.",
        "operationId" : "diffVersions",
        "parameters" : [ {
          "description" : "Ontology acronym.",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Base version (version_id or tag).",
          "in" : "query",
          "name" : "from",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Target version (version_id or tag, e.g. latest).",
          "in" : "query",
          "name" : "to",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VersionDiff"
                }
              }
            },
            "description" : "The vocabulary diff between the two versions"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Ontology or version not found locally"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Diff two local versions of an ontology",
        "tags" : [ "Ontologies" ]
      }
    },
...
    "/bioportal/ontologies/{ontology}/classes/roots" : {
      "get" : {
        "description" : "Get root classes in a particular ontology. For the CEDARPC ontology, all provisional classes in it will be returned.",
        "operationId" : "findRootClasses",
        "parameters" : [ {
          "description" : "BioPortal ontology identifier. Examples: NCIT, FMA, OBI.",
          "in" : "path",
          "name" : "ontology",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/OntologyClass"
                  }
                }
              }
            },
            "description" : "The ontology's root classes"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Bad request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Not found"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Get root classes",
        "tags" : [ "Classes", "Ontologies" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Ontology" : {
        "type" : "object",
        "properties" : {
          "@id" : {
            "type" : "string"
          },
          "@type" : {
            "type" : "string"
          },
          "details" : {
            "$ref" : "#/components/schemas/OntologyDetails"
          },
          "flat" : {
            "type" : "boolean"
          },
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          }
        }
      },
      "OntologyClass" : {
        "type" : "object",
        "properties" : {
          "@id" : {
            "type" : "string"
          },
          "@type" : {
            "type" : "string"
          },
          "created" : {
            "type" : "string"
          },
          "creator" : {
            "type" : "string"
          },
          "definitions" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "hasChildren" : {
            "type" : "boolean"
          },
          "id" : {
            "type" : "string"
          },
          "ontology" : {
            "type" : "string"
          },
          "prefLabel" : {
            "type" : "string"
          },
          "provisional" : {
            "type" : "boolean"
          },
          "relations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Relation"
            }
          },
          "subclassOf" : {
            "type" : "string"
          },
          "synonyms" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "type" : {
            "type" : "string"
          }
        }
      },
...
      "OntologyVersion" : {
        "type" : "object",
        "properties" : {
          "effectiveDate" : {
            "type" : "string"
          },
          "latest" : {
            "type" : "boolean"
          },
          "released" : {
            "type" : "string"
          },
          "version" : {
            "type" : "string"
          },
          "versionId" : {
            "type" : "string"
          }
        }
      },
...
      "Relation" : {
        "type" : "object",
        "properties" : {
          "@id" : {
            "type" : "string"
          },
          "@type" : {
            "type" : "string"
          },
          "created" : {
            "type" : "string"
          },
          "creator" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "relationType" : {
            "type" : "string"
          },
          "sourceClassId" : {
            "type" : "string"
          },
          "targetClassId" : {
            "type" : "string"
          },
          "targetClassOntology" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          }
        }
      },
...
      "VersionTriple" : {
        "type" : "object",
        "properties" : {
          "declaredVersion" : {
            "type" : "string"
          },
          "effectiveDate" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          }
        }
      }
    },
    "securitySchemes" : {
      "api_key" : {
        "description" : "Your CEDAR API key, using the format 'apiKey <yourApiKey>'",
        "in" : "header",
        "name" : "Authorization",
        "type" : "apiKey"
      }
    }
  }
}